requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0
